requirements.txt
pyrogram
tgcrypto
pytgcalls
yt-dlp
gtts
openai
python-dotenv
